requirements.txt
futu-api
pandas
numpy
pyyaml
pytest
typing-extensions 